+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Comparisons support for tuples.
 - Standalone versions of methods: `pw.groupby`, `pw.join`, `pw.join_inner`, `pw.join_left`, `pw.join_right` and `pw.join_outer`.
 - The ability to use python's `abs` function on Pathway expressions.
+- `asof_join` now has configurable temporal behavior (delaying outputs, ignoring late entries and cleaning unused memory). The configuration can be passed using the `behavior` parameter of `asof_join` method.
 
 ### Changed
 - `interval_join` can now also work with intervals of zero length.